修改bug

+++ Finance/constant/SHGBusinessMargin.h
@@ -18,8 +18,8 @@
 #define kButtonTopMargin MarginFactor(13.0f)
 
 #define kFourButtonWidth floorf((SCREENWIDTH - 2 * MarginFactor(12.0f) - 3 * MarginFactor(10.0f)) / 4.0)
-#define kThreeButtonWidth floor((SCREENWIDTH - 2 * MarginFactor(12.0f) - 2 * MarginFactor(10.0f)) / 3.0)
-#define kTwoButtonWidth floor((SCREENWIDTH - 2 * MarginFactor(12.0f) -MarginFactor(10.0f)) / 2.0)
+#define kThreeButtonWidth floorf((SCREENWIDTH - 2 * MarginFactor(12.0f) - 2 * MarginFactor(10.0f)) / 3.0)
+#define kTwoButtonWidth floorf((SCREENWIDTH - 2 * MarginFactor(12.0f) -MarginFactor(10.0f)) / 2.0)
 
 //定位界面
 #define kLocationButtonLeftMargin floorf((SCREENWIDTH - 3 * MarginFactor(111.0f) - 2 * MarginFactor(12.0f)) / 2.0f)
